added echos

+++ docker/install-packer.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
-yum -y install unzip
+if [ ! -f /usr/bin/unzip ]; then
+  echo "Downloadin upzip..."
+  yum -y install unzip
+fi
 
 #Installing Packer
-wget -q https://dl.bintray.com/mitchellh/packer/packer_0.8.6_linux_amd64.zip
-mkdir -p /opt/packer; unzip -o packer_0.8.6_linux_amd64.zip -d /opt/packer
+if [ ! -f /opt/packer/packer ]; then
+  echo "Downloading Packer (126MB)..."
+  wget -q https://dl.bintray.com/mitchellh/packer/packer_0.8.6_linux_amd64.zip
+  mkdir -p /opt/packer; unzip -o packer_0.8.6_linux_amd64.zip -d /opt/packer
+fi
